reorganize style module
