+++ src/style/layers/mod.rs
@@ -1,0 +1,11 @@
+mod circle;
+mod fill;
+mod id;
+mod line;
+mod symbol;
+
+pub use circle::CircleLayer;
+pub use fill::FillLayer;
+pub use id::LayerId;
+pub use line::{LineCap, LineJoin, LineLayer};
+pub use symbol::{SymbolAnchor, SymbolLayer};
