+++ src/style/sources/mod.rs
@@ -1,0 +1,5 @@
+mod geojson_source;
+mod id;
+
+pub use geojson_source::GeoJsonSource;
+pub use id::SourceId;
